ALPH_VA/requirement.txt
httpx
setuptools
pyttsx3
google-genai
voicevox-client
googletrans
TTS
configparser
SpeechRecognition
PyAudio